Bluebird Combinator
This is getting it's own branch because, for the moment, it requires
unstable features. I might rename this branch to `unstable` at some
point, but for now it's just this one combinator that requires it so I'm
going to limit it.

Not too much to talk about here. I wrote a auto-curry macro, which also
requires unstable features. Once I get generics working with it, I'll
rewrite this to use the macro, in the mean time, it's sorta ugly, but it
does the job and I think it does the job in the best/only way possible.